// interfaces/flight.interface.ts
export interface IFlights {
	result: Result;
}

export interface Result {
	flights: FlightElement[];
	bestPrices: BestPrices;
}

export interface BestPrices {
	ONE_CONNECTION: Direct;
	DIRECT: Direct;
}
// ...
export interface FlightElement {
	hasExtendedFare: boolean;
	flight: FlightFlight;
	flightToken: string;
}